code/requirements.txt
groq>=0.12.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
